backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
krakenex==2.2.1
openai==1.3.5
requests==2.31.0
pandas==2.1.3
ta==0.11.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
numpy==1.26.2
apscheduler==3.10.4


